Create reusable hooks and components
