+++ aries-site/src/examples/templates/dashboards/components/index.js
@@ -10,6 +10,7 @@ export * from './ItemCountList';
 export * from './Legend';
 export * from './Measure';
 export * from './MeterGroup';
+export * from './MonthlySpend';
 export * from './Greeting';
 export * from './PageHeader';
 export * from './StatusBar';
