+++ aries-site/src/examples/templates/dashboards/content/chartCards/utils/index.js
@@ -1,3 +1,4 @@
 export * from './formatValues';
+export * from './hooks';
 export * from './window';
 export * from './statistics';
